src/cli/pyproject.toml
[project]
name = "doremix-cli"
version = "0.0.1"
description = "CLI version of DORéMiX"
dependencies = [
    "dotenv>=0.9.9",
    "pydantic>=2.12.5",
    "python-vlc>=3.0.21203",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "typer>=0.21.0",
    "yt-dlp>=2025.12.8",
]
